requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<3.14
pandas>=1.1.0
matplotlib>=3.1.3
